+++ src/Graphic/Batch.cpp
@@ -231,7 +231,7 @@ void Batch::draw(const mat3f& transform, Text&& text)
 		uint32_t c = utf8::next(start, end);
 		// TODO check if rendering text out of screen for culling ?
 		const Character& ch = text.font->getCharacter(c);
-		vec2f position = vec2f(advance + ch.bearing.x, -(ch.size.y - ch.bearing.y)) * scale;
+		vec2f position = vec2f(advance + ch.bearing.x, (float)-(ch.size.y - ch.bearing.y)) * scale;
 		vec2f size = vec2f((float)ch.size.x, (float)ch.size.y) * scale;
 		draw(transform, Batch::Rect(position, size, ch.texture.get(0), ch.texture.get(1), ch.texture.texture, text.color, text.layer));
 		// now advance cursors for next glyph (note that advance is number of 1/64 pixels)
